backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
sqlalchemy==2.0.32
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.4.0
pydantic[email]==2.8.2
psycopg2-binary==2.9.9
bcrypt==4.0.1
